Fix: Recover scroll position when CNE form is redraw
If the value of some fields of the CNE form changes, the form must be redrawn.
When the form is redrawn, the scroll position is reset to zero.
Maintain scroll position when redrawing form by keeping scrollTop value first and recovering it after redrawing.

This PR is based on #67 PR.